dev_requirements.txt
-r requirements.txt

django-print-sql==2018.3.6
sqlparse==0.3.1
pytest-django==4.2.0
lxml==4.9.1
coverage==5.5
pytest-lazy-fixture==0.6.3
django-debug-toolbar==3.7.0
pytest-mock==3.10.0
